Move env_Var function to stop circular deps
