+++ pkgs/flojoy/flojoy/__init__.py
@@ -17,3 +17,4 @@ from .node_preflight import *  # noqa: F403
 from .instruments import *  # noqa: F403
 from .models import *  # noqa: F403
 from .connection_manager import *  # noqa: F403
+from .env_var import *  # noqa: F403
