+++ pkgs/flojoy/flojoy/utils.py
@@ -1,9 +1,8 @@
 import decimal
 import json as _json
 import os
-import sys
 from pathlib import Path
-from typing import Any, Callable, Optional
+from typing import Any, Callable
 
 import logging
 import numpy as np
@@ -18,19 +17,14 @@ from huggingface_hub import snapshot_download
 from flojoy.connection_manager import DeviceConnectionManager
 from .dao import Dao
 from .config import FlojoyConfig, logger
-
+from .env_var import get_env_var
 from .node_init import NodeInit, NodeInitService
 
-import keyring
-from keyrings.cryptfile.cryptfile import CryptFileKeyring
 import base64
-from .CONSTANTS import FLOJOY_DIR, FLOJOY_CACHE_DIR, CREDENTIAL_FILE, KEYRING_KEY
+from .CONSTANTS import FLOJOY_DIR, FLOJOY_CACHE_DIR, CREDENTIAL_FILE
 
 
 __all__ = [
-    "get_env_var",
-    "set_env_var",
-    "delete_env_var",
     "get_credentials",
     "hf_hub_download",
     "snapshot_download",
@@ -304,64 +298,6 @@ def get_flojoy_root_dir() -> str:
         root_dir = yaml_dict["PATH"]
 
     return root_dir
-
-
-def get_keyring():
-    if sys.platform.lower() == "linux":
-        kr = CryptFileKeyring()
-        kr.keyring_key = KEYRING_KEY
-        keyring.set_keyring(kr)
-    return keyring
-
-
-def get_env_var(key: str) -> Optional[str]:
-    kr = get_keyring()
-    return kr.get_password("flojoy", key)
-
-
-def set_env_var(key: str, value: str):
-    kr = get_keyring()
-    kr.set_password("flojoy", key, value)
-    home = str(Path.home())
-    file_path = os.path.join(home, os.path.join(FLOJOY_DIR, CREDENTIAL_FILE))
-
-    if not os.path.exists(file_path):
-        logger.info(f"{file_path} does not exist")
-        with open(file_path, "w") as f:
-            f.write(key)
-        logger.info(f"Env var written to {file_path}")
-        return
-
-    logger.info(f"{file_path} exists, writing env to {file_path}")
-    with open(file_path, "r") as f:
-        keys = f.read().strip().split(",")
-        if key not in keys:
-            keys.append(key)
-
-    with open(file_path, "w") as f:
-        f.write(",".join(keys))
-
-
-def delete_env_var(key: str):
-    home = str(Path.home())
-    file_path = os.path.join(home, os.path.join(FLOJOY_DIR, CREDENTIAL_FILE))
-
-    if not os.path.exists(file_path):
-        return
-
-    with open(file_path, "r") as f:
-        keys = f.read().strip().split(",")
-
-    if key not in keys:
-        return
-
-    keys.remove(key)
-
-    with open(file_path, "w") as f:
-        f.write(",".join(keys))
-
-    kr = get_keyring()
-    kr.delete_password("flojoy", key)
 
 
 def get_credentials() -> list[dict[str, str]]:
